+++ src/sisl/io/wannier90/seedname.py
@@ -4,11 +4,14 @@
 """
 Sile object for reading/writing Wannier90 in/output
 """
+from collections import defaultdict
 from pathlib import Path
 
 import numpy as np
+import scipy.sparse as sps
 from scipy.sparse import lil_matrix
 
+import sisl._array as _a
 from sisl import Geometry, Lattice
 from sisl.messages import deprecate_argument
 from sisl.physics import Hamiltonian
@@ -20,6 +23,27 @@ from ..sile import *
 from .sile import SileWannier90
 
 __all__ = ["winSileWannier90"]
+
+
+def _construct_hamiltonian(geometry: Geometry, Hsc):
+    # Get supercell size
+    nsc = _a.zerosi(3)
+    for isc in Hsc.keys():
+        for i in (0, 1, 2):
+            nsc[i] = max(abs(isc[i]), nsc[i])
+
+    # Create the full supercell
+    nsc = nsc * 2 + 1
+    geometry.set_nsc(nsc)
+
+    # Create the big matrix
+    H = []
+    for isc in geometry.lattice.sc_off:
+        H.append(Hsc[tuple(isc)].tocsr())
+
+    H = sps.hstack(H)
+
+    return Hamiltonian.fromsp(geometry, H)
 
 
 class winSileWannier90(SileWannier90):
@@ -259,12 +283,23 @@ class winSileWannier90(SileWannier90):
         self._set_file()
         self._write_geometry(geom, fmt, *args, **kwargs)
 
-    @sile_fh_open()
-    def _r_hamiltonian(self, geom, dtype=np.float64, **kwargs):
-        """Reads a Hamiltonian
+    def _r_wigner_seitz_weights(self):
+        # Number of Wigner-Seitz degeneracy points
+        npts = int(self.readline())
 
-        Reads the Hamiltonian model
-        """
+        ws = []
+        while len(ws) < npts:
+            # both formats uses 15 points per line,
+            # however, this should be usable if they decide to change
+            # the number of counts per line.
+            ws.extend(list(map(int, self.readline().split())))
+
+        ws = 1.0 / _a.arrayd(ws)
+        return ws
+
+    @sile_fh_open()
+    def _r_hamiltonian_hr(self, geometry, dtype=np.float64, **kwargs):
+        """Reads a Hamiltonian model from the _hr.dat file"""
         cutoff = kwargs.get("cutoff", 0.00001)
 
         # Rewind to ensure we can read the entire matrix structure
@@ -275,38 +310,21 @@ class winSileWannier90(SileWannier90):
 
         # Number of orbitals
         no = int(self.readline())
-        if no != geom.no:
+        if no != geometry.no:
             raise ValueError(
-                self.__class__.__name__
-                + ".read_hamiltonian has found inconsistent number "
+                f"{self.__class__.__name__}"
+                ".read_hamiltonian has found inconsistent number "
                 "of orbitals in _hr.dat vs the geometry. Remember to re-run Wannier90?"
             )
 
-        # Number of Wigner-Seitz degeneracy points
-        nrpts = int(self.readline())
-
-        # First read across the Wigner-Seitz degeneracy
-        # This is formatted with 15 per-line.
-        if nrpts % 15 == 0:
-            nlines = nrpts
-        else:
-            nlines = nrpts + 15 - nrpts % 15
-
-        ws = []
-        for _ in range(nlines // 15):
-            ws.extend(list(map(int, self.readline().split())))
-
-        # Convert to numpy array and invert (for weights)
-        ws = 1.0 / np.array(ws, np.float64).flatten()
-
-        # Figure out the number of supercells
-        # and maintain the Hamiltonian in the ham list
-        nsc = [0, 0, 0]
+        ws = self._r_wigner_seitz_weights()
 
         # List for holding the Hamiltonian
-        ham = []
-        iws = -1
+        Hsc = defaultdict(lambda: lil_matrix((geometry.no, geometry.no), dtype=dtype))
+        is_complex = np.iscomplexobj(dtype(1))
 
+        iws = -1
+        isc = [0, 0, 0]
         while True:
             l = self.readline()
             if l == "":
@@ -316,72 +334,123 @@ class winSileWannier90(SileWannier90):
             l = l.split()
 
             # Get super-cell, row and column
-            iA, iB, iC, r, c = map(int, l[:5])
-
-            nsc[0] = max(nsc[0], abs(iA))
-            nsc[1] = max(nsc[1], abs(iB))
-            nsc[2] = max(nsc[2], abs(iC))
+            isc[0], isc[1], isc[2], r, c = map(int, l[:5])
 
             # Update index for degeneracy, if required
             if r + c == 2:
                 iws += 1
 
             # Get degeneracy of this element
-            f = ws[iws]
+            w = ws[iws]
 
-            # Store in the Hamiltonian array:
-            #   isc
-            #   row
-            #   column
-            #   Hr
-            #   Hi
-            ham.append(([iA, iB, iC], r - 1, c - 1, float(l[5]) * f, float(l[6]) * f))
+            # Scale matrix elements
+            hr = float(l[5]) * w
+            if is_complex:
+                h = hr + 1j * float(l[6]) * w
+            else:
+                h = hr
 
-        # Update number of super-cells
-        geom.set_nsc([i * 2 + 1 for i in nsc])
+            if abs(h) > cutoff:
+                Hsc[tuple(isc)][r - 1, c - 1] = h
 
-        # With the geometry in place we can read in the entire matrix
-        # Create a new sparse matrix
-        Hr = lil_matrix((geom.no, geom.no_s), dtype=dtype)
-        Hi = lil_matrix((geom.no, geom.no_s), dtype=dtype)
+        return _construct_hamiltonian(geometry, Hsc)
 
-        # populate the Hamiltonian by examining the cutoff value
-        for isc, r, c, hr, hi in ham:
-            # Calculate the column corresponding to the
-            # correct super-cell
-            c = c + geom.sc_index(isc) * geom.no
+    @sile_fh_open()
+    def _r_hamiltonian_tb(self, geometry, dtype=np.float64, **kwargs):
+        """Reads a Hamiltonian model from the _tb.dat file"""
+        cutoff = kwargs.get("cutoff", 0.00001)
 
-            if abs(hr) > cutoff:
-                Hr[r, c] = hr
-            if abs(hi) > cutoff:
-                Hi[r, c] = hi
-        del ham
+        # Rewind to ensure we can read the entire matrix structure
+        self.fh.seek(0)
 
-        if np.dtype(dtype).kind == "c":
-            Hr = Hr.tocsr()
-            Hi = Hi.tocsr()
-            Hr = Hr + 1j * Hi
+        # Time of creation
+        self.readline()
 
-        return Hamiltonian.fromsp(geom, Hr)
+        #  Lattice vectors [Ang]
+        for _ in range(3):
+            self.readline()
+        # TODO check against self.lattice?!
 
-    def read_hamiltonian(self, *args, **kwargs):
-        """Read the electronic structure of the Wannier90 output
+        # Number of orbitals
+        no = int(self.readline())
+        if no != geometry.no:
+            raise ValueError(
+                f"{self.__class__.__name__}"
+                ".read_hamiltonian has found inconsistent number "
+                "of orbitals in _hr.dat vs the geometry. Remember to re-run Wannier90?"
+            )
+
+        ws = self._r_wigner_seitz_weights()
+
+        # List for holding the Hamiltonian
+        Hsc = defaultdict(lambda: lil_matrix((geometry.no, geometry.no), dtype=dtype))
+        is_complex = np.iscomplexobj(dtype(1))
+
+        # Parse hamiltonian matrix elements
+        for w in ws:
+            l = self.readline()  # Skip empty line
+            if not l.strip() == "":
+                raise ValueError(
+                    f"{self.__class__.__name__}"
+                    ".read_hamiltonian unable to parse <>_tb.dat file, due to "
+                    "error in file format."
+                )
+
+            # Get super-cell
+            isc = map(int, self.readline().split())
+
+            # Get Hamiltonian matrix elements
+            Hr = Hsc[tuple(isc)]
+            for _ in range(no**2):
+                l = self.readline().split()
+
+                # Get row and column):
+                r, c = map(int, l[:2])
+
+                # Scale matrix elements
+                hr = float(l[2]) * w
+                if is_complex:
+                    h = hr + 1j * float(l[3]) * w
+                else:
+                    h = hr
+
+                if abs(h) > cutoff:
+                    Hr[r - 1, c - 1] = h
+
+        return _construct_hamiltonian(geometry, Hsc)
+
+    def read_hamiltonian(self, cutoff: float = 1e-5, *args, **kwargs):
+        """Read the electronic structure of the Wannier90 output by reading the <>_tb.dat, <>_hr.dat
 
         Parameters
         ----------
-        cutoff: float, optional
+        cutoff:
            the cutoff value for the zero Hamiltonian elements, default
            to 0.00001 eV.
+
+        dtype: np.float64, optional
+            the default data-type used for the matrix.
+            Is mainly useful to check whether the TB model has imaginary
+            components (it should not since it is a Wannier model).
         """
         # Retrieve the geometry...
         geom = self.read_geometry()
 
-        # Set file
-        self._set_file("_hr.dat")
+        order = kwargs.pop("order", ["tb", "hr"])
+        for f in order:
+            # Set file
+            self._set_file(f"_{f.lower()}.dat")
 
-        H = self._r_hamiltonian(geom, *args, **kwargs)
-        self._set_file()
-        return H
+            # only parse if the file exists
+            if self._file.exists():
+                H = getattr(self, f"_r_hamiltonian_{f.lower()}")(
+                    geom, *args, cutoff=cutoff, **kwargs
+                )
+            self._set_file()
+            if H is not None:
+                return H
+
+        return None
 
     def ArgumentParser(self, p=None, *args, **kwargs):
         """Returns the arguments that is available for this Sile"""